requirements.txt
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iop-sdk-python @ file://ss
pyaes==1.6.1
pyasn1==0.6.2
python-dotenv==1.2.1
python-telegram-bot==22.6
requests==2.32.5
rsa==4.9.1
Telethon==1.42.0
typing_extensions==4.15.0
urllib3==2.6.3
